Updated to beta1. Changed items include:
o Got rid of module as it is cvs-specific; it's been replaced by rootDirectory (synomous with module in CVS) and baseDirectory which is the greatest common subdirectory of the commit; this is used for the matching in the config file.

o Added directory/indirectlyAffectedDirectories to the Model.

o config/Controller.pm now does matching against baseDirectory in determining how to setup view instances instead of a strict match on module/rootDirectory.