Patch by Michael Haggerty (haggerty@jpk.com):
This first patch causes cvs to always be invoked with the "-f" option, which causes it to skip reading the user's .cvsrc file.  Without this option, non-standard options in the users .cvsrc (for example, a different choice of diff options or even an external diff program) would cause arbitrary problems.