Patch by Michael Haggerty (haggerty@jpk.com):
The last remaining acceptance test failure was due to inconsistent handling of missing trailing newline characters in files.  Since it is conventional that files end with a newline, I changed the test code to add the newline, and changed the test case expectations correspondingly.

I also added a Commit.__str__() method that includes the commit message.
 This makes it easier to find the source of test failures.